add @waku/react, use hook from there
