+++ examples/web-chat/package.json
@@ -7,11 +7,12 @@
     "@libp2p/bootstrap": "^5.0.0",
     "@livechat/ui-kit": "^0.5.0-24",
     "@multiformats/multiaddr": "11.0.7",
+    "@waku/react": "file:../../../waku-ui",
     "@waku/byte-utils": "^0.0.2",
     "@waku/core": "^0.0.10",
     "@waku/create": "^0.0.4",
     "@waku/dns-discovery": "0.0.5",
-    "@waku/interfaces": "^0.0.5",
+    "@waku/interfaces": "^0.0.7",
     "@waku/peer-exchange": "^0.0.3",
     "process": "^0.11.10",
     "protons-runtime": "^3.1.0",
